requirements.txt
python-telegram-bot==21.7
supabase==2.10.0
python-dotenv==1.0.1
httpx==0.27.2
flask==3.0.0
google-generativeai>=0.8.0
pypdf>=4.0.0
apscheduler>=3.10.0
